Add a function to search find a command given a partial name.
When given a command (including the path) that does not have a file
extension, use the PATHEXT environment variable to find the command.
This is the same behavior used by the command shell to find commands in
the PATH.  See https://technet.microsoft.com/en-us/library/cc723564.aspx
for more information.